requirements.txt
asgiref==3.8.1
astroid==3.3.8
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
diff-match-patch==20241021
dill==0.3.9
django
django-debug-toolbar==5.1.0
django-filter==25.1
django-import-export==4.3.7
django-jet-reboot==1.3.10
django-tinymce==4.1.0
djangorestframework==3.16.0
feedparser==6.0.11
google-api-python-client==1.4.1
httplib2==0.22.0
idna==3.10
isort==6.0.0
mailtrap==2.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
sgmllib3k==1.0.0
six==1.17.0
slugify==0.0.1
sqlparse==0.5.3
tablib==3.8.0
tomlkit==0.13.2
uritemplate==4.1.1
urllib3==2.4.0
